+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/commands/HangCommand.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.common.commands;
+/*package org.firstinspires.ftc.teamcode.common.commands;
 
 import com.arcrobotics.ftclib.command.CommandBase;
 import com.arcrobotics.ftclib.command.InstantCommand;
@@ -35,3 +35,4 @@ public class HangCommand extends InstantCommand {
         return Math.abs(hangSubsystem.getError()) < 10;
     }
 }*/
+
