+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/commands/intake/IntakeRotatorCommand.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.common.commands;
+package org.firstinspires.ftc.teamcode.common.commands.intake;
 
 import com.arcrobotics.ftclib.command.InstantCommand;
