+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/robot/HardwareMapNames.java
@@ -50,14 +50,9 @@ public class HardwareMapNames {
 
     // Deposit
     public static String BUCKET_SERVO = "bucket";
-    public static String CLAW_SERVO = "claw";
-
-    // Specimen
-    public static String SPECIMEN_SERVO = "specimen";
 
     // Hang
     public static String HANG_MOTOR_1 = "hang1";
-    public static String HANG_MOTOR_2 = "hang2";
 
     // Odometry
     public static String OTOS = "sensor_otos";
