+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/test/subsystem_tests/HangCommandTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.test.subsystem_tests;
+/*package org.firstinspires.ftc.teamcode.test.subsystem_tests;
 
 import com.arcrobotics.ftclib.command.CommandOpMode;
 import com.arcrobotics.ftclib.command.CommandScheduler;
@@ -40,3 +40,5 @@ public class HangCommandTest extends CommandOpMode {
         CommandScheduler.getInstance().run();
     }
 }
+
+ */
